Test: Fix broken test
Aligning to the right is done regardless of the number of boxes

Therefore if there is only one box, it might align the right border
of the box with the box underneath, so that the next slice is free.
Aligning on the left of the shelf might mean that the 2nd slice can not
be used.